Move datatypes code from Translator.cs Translator.Datatypes.cs
